requirements.txt
amqp==2.6.1
asgiref==3.5.2
beautifulsoup4==4.11.1
billiard==3.6.4.0
bs4==0.0.1
celery==4.3.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==4.1.2
django-admin-tailwind==0.0.16
django-compat==1.0.15
djangorestframework==3.14.0
drf-yasg==1.21.4
html-parser==0.2
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
kombu==4.6.11
lxml==4.9.1
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
Pillow==9.2.0
ply==3.11
psycopg2-binary==2.9.5
pyparsing==3.0.9
pytz==2022.5
redis==3.2.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.2.post1
sqlparse==0.4.3
tabulate==0.9.0
uritemplate==4.1.1
urllib3==1.26.12
vine==1.3.0
